Make undo work
User need to manually update the component prop to tell when to flush
the editor with new content, i.e when the editor should discard the undo
and set the content of the editor to provided content.

The editor by default do not actually update the DOM content for React
component. User sees the content getting updated, content udpates in the
state, but it's not updated for React component. This is done because we
now take the editor's content from props; so every time new props are
received, new component is rendered, and old "undo" information is
lost. The alternative was to implement a custom undo functionality,
which I feel is more of a burden then making user flush the content to
DOM manually. User can simply pass `flushEditorDOM=true` to now use undo
functionality and update DOM content as the props.